Date:        06.03.2016 Author:      Kasimov Ildar Description:
Changes to be committed:
	modified:   .vs/SimulationOfBusRoute/v14/.suo
	modified:   SimulationOfBusRoute/MainForm.Designer.cs
	modified:   SimulationOfBusRoute/MainForm.cs
	modified:   SimulationOfBusRoute/Models/CBusRoute.cs
	modified:   SimulationOfBusRoute/Models/CBusStation.cs
	new file:   SimulationOfBusRoute/Presenters/CBasePresenter.cs
	modified:   SimulationOfBusRoute/Presenters/CMainFormPresenter.cs
	modified:   SimulationOfBusRoute/SimulationOfBusRoute.csproj
	modified:   SimulationOfBusRoute/Views/IMainFormView.cs
	modified:   SimulationOfBusRoute/bin/Debug/SimulationOfBusRoute.exe
	modified:   SimulationOfBusRoute/bin/Debug/SimulationOfBusRoute.pdb
	new file:   SimulationOfBusRoute/bin/Debug/SimulationOfBusRoute.vshost.exe.manifest
	modified:   SimulationOfBusRoute/obj/Debug/SimulationOfBusRoute.csproj.GenerateResource.Cache
	modified:   SimulationOfBusRoute/obj/Debug/SimulationOfBusRoute.exe
	modified:   SimulationOfBusRoute/obj/Debug/SimulationOfBusRoute.pdb

+++ SimulationOfBusRoute/Models/CBusRoute.cs
@@ -58,7 +58,7 @@ namespace SimulationOfBusRoute.Models
             mBuses.Add(new CBus(newBusID, startStationIndex, mBusStations[(int)startStationIndex].Position, maxNumOfPassengers));
         }
 
-        public void AddBusStation(TPoint2 position, ushort startNumOfPassengers, uint intensity)
+        public void AddBusStation(TPoint2 position, ushort startNumOfPassengers, ushort intensity)
         {
             uint newBusStationID = Convert.ToUInt32(mBusStations.Count + 1);
